front: split NGE-to-OSRD trainrun-related logic in separate file
